+++ tests/static_index_bounds_invalid.ml
@@ -1,0 +1,6 @@
+function reproduce()
+  values = [1, 2, 3]
+  print values[3]
+end function
+
+reproduce()