+++ tests/static_index_type_invalid.ml
@@ -1,0 +1,7 @@
+function reproduce()
+  values = [1, 2, 3]
+  badIndex = 1.5
+  print values[badIndex]
+end function
+
+reproduce()