+++ tests/static_struct_member_invalid.ml
@@ -1,0 +1,10 @@
+struct Player
+  name
+end struct
+
+function reproduce()
+  player = Player("Ada")
+  print player.missing
+end function
+
+reproduce()